requirements.txt
amqp==5.1.1
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
autopep8==1.7.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.11
django-appconf==1.0.5
django-autoslug==1.9.8
django-celery-email==3.0.0
django-cors-headers==3.10.1
django-countries==7.2.1
django-environ==0.8.1
django-filter==21.1
django-haystack==3.1.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-haystack==1.8.11
drf-yasg==1.20.0
flake8==4.0.1
flower==1.0.0
humanize==4.3.0
idna==3.3
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.10.1
phonenumbers==8.12.33
Pillow==9.0.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
tornado==6.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
Whoosh==2.7.4
wrapt==1.14.1
